+++ src/class/Admin/Admin.php
@@ -170,31 +170,35 @@ final class Admin {
 	 * Register and enqueue gutenberg blocks, must be executed on the whole website
 	 *
 	 * @since 4.1 Using block.json, added script translation, added lumiere_scripts_admin_gutenberg script
-	 * @see \Lumiere\Admin\Widget_Selection::lumiere_register_widget_block() Register widget block, not included here
+	 * @since 4.6.6 Use wp_register_block_types_from_metadata_collection() if file exists, needs WordPress 6.8
+	 *
+	 * @see \Lumiere\Admin\Widget_Selection::lumiere_register_widget_block() Widget block registered there, not included here
 	 */
 	public function lum_enqueue_blocks(): void {
 
 		$block_dir = LUM_WP_PATH . 'assets/blocks';
 		$blocks = [ 'post', 'addlink', 'opensearch', 'widget-sidebar-options' ];
 		$extra_render = [];
-		foreach ( $blocks as $block ) {
 
-			// Include the render_callback if render.php exists
-			/*
-			$file_render = LUM_WP_PATH . 'assets/blocks/' . $block . '/render.php';
-			if ( is_file( $file_render ) ) {
-				require_once $file_render;
-				$callback = 'lum_render_block_' . $block;
-				$extra_render = [ 'render_callback' => $callback ];
-			}*/
+		if ( function_exists( 'wp_register_block_types_from_metadata_collection' ) && file_exists( Get_Options::LUM_BLOCKS_MANIFEST ) ) {
+			wp_register_block_types_from_metadata_collection(
+				$block_dir,
+				Get_Options::LUM_BLOCKS_MANIFEST
+			);
+		} elseif ( function_exists( 'register_block_type' ) ) {
+			foreach ( $blocks as $block ) {
 
-			register_block_type( $block_dir . '/' . $block, $extra_render );
-			/*add_action( => using block.json and "wp i8n" generation
-				'init',
-				function( string $block ) {
-					wp_set_script_translations( 'lumiere-movies-' . $block . '-editor-script', 'lumiere-movies', LUM_WP_PATH . 'languages/' );
-				}
-			);*/
+				// Include the render_callback if render.php exists
+				/*
+				$file_render = LUM_WP_PATH . 'assets/blocks/' . $block . '/render.php';
+				if ( is_file( $file_render ) ) {
+					require_once $file_render;
+					$callback = 'lum_render_block_' . $block;
+					$extra_render = [ 'render_callback' => $callback ];
+				}*/
+
+				register_block_type( $block_dir . '/' . $block, $extra_render );
+			}
 		}
 
 		// Javascripts functions for Gutenberg blocks.